\name{loadConfig}
\alias{loadConfig}
\title{loadConfig}
\description{Loading configuration from file}
\usage{loadConfig(fileNames = c("config.yml"), ...)}
\arguments{
  \item{fileNames}{character vector: fileNames, order matters: duplicate entries are overwritten}
  \item{\dots}{further arguments passed to new("twConfig")}
}



\author{Thomas Wutzler}





