\name{loadYaml}
\alias{loadYaml}
\title{loadYaml}
\description{parsing and Yaml file into a configuration}
\usage{loadYaml(object, fileName = "config.yml", ...)}
\arguments{
  \item{object}{the class object}
  \item{fileName}{the file to parse}
  \item{\dots}{
}
}



\author{Thomas Wutzler}





