\name{twConfig-package}
\alias{twConfig-package}
\alias{twConfig}
\docType{package}
\title{Cascading configuration files}
\description{Cascading configuration files.}
\details{
\tabular{ll}{Package: \tab twConfig\cr
Title: \tab Cascading configuration files\cr
Version: \tab 0.8.2\cr
Date: \tab 2012-02-28\cr
Author: \tab Thomas Wutzler\cr
Maintainer: \tab Thomas Wutzler <twutz@bgc-jena.mpg.de>\cr
License: \tab Apache License (== 2.0)\cr
LazyLoad: \tab yes\cr
Depends: \tab R (>= 1.12.0)\cr
Imports: \tab yaml\cr
Suggests: \tab RUnit\cr}
}
\author{Thomas Wutzler}

\keyword{ package }


